Add abstract Contact class of which Address, Email, and Phone class extends
